OPENEUROPA-2340: Add output of site specific footer for EU footer.
