+++ oe_corporate_blocks.module
@@ -21,7 +21,7 @@ function oe_corporate_blocks_theme() {
     'oe_corporate_blocks_eu_footer' => [
       'variables' => [
         'corporate_footer' => [],
-        'custom_footer' => [],
+        'site_specific_footer' => [],
       ],
     ],
   ];
