+++ src/Plugin/Block/EcFooterBlock.php
@@ -7,7 +7,6 @@ namespace Drupal\oe_corporate_blocks\Plugin\Block;
 use Drupal\Component\Utility\NestedArray;
 use Drupal\Core\Cache\CacheableMetadata;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Drupal\oe_corporate_blocks\Entity\FooterLinkSocialInterface;
 
 /**
  * Provides the corporate footer block.
@@ -43,59 +42,11 @@ class EcFooterBlock extends FooterBlockBase implements ContainerFactoryPluginInt
 
     NestedArray::setValue($build, ['#corporate_footer', 'bottom_links'], $config->get('bottom_links'));
 
-    $general_links = $this->getSiteSpecificFooterLinks('footer_link_general', $cache);
-    $social_links = $this->getSiteSpecificFooterLinks('footer_link_social', $cache);
-    $site_info_config = $this->configFactory->get('system.site');
-    $cache->addCacheableDependency($site_info_config);
-    $site_identity = $site_info_config->get('name');
-
-    if (!empty($social_links) || !empty($general_links)) {
-      NestedArray::setValue($build, ['#site_specific_footer', 'site_identity'], $site_identity);
-      NestedArray::setValue($build, ['#site_specific_footer', 'social_links'], $social_links);
-      NestedArray::setValue($build, ['#site_specific_footer', 'other_links'], $general_links);
-    }
+    $this->setSiteSpecificFooter($build, $cache);
 
     $cache->applyTo($build);
 
     return $build;
-  }
-
-  /**
-   * Retrieve a list of custom footer link configuration entities.
-   *
-   * @param string $type
-   *   Type of configs.
-   * @param \Drupal\Core\Cache\CacheableMetadata $cache
-   *   The Cacheable metadata abject.
-   *
-   * @return array
-   *   The array of links.
-   */
-  protected function getSiteSpecificFooterLinks(string $type, CacheableMetadata &$cache): array {
-    $links = [];
-    /** @var \Drupal\Core\Entity\EntityStorageInterface $links_storage */
-    $links_storage = $this->entityTypeManager->getStorage($type);
-    $cache->addCacheTags($links_storage->getEntityType()->getListCacheTags());
-    $link_ids = $links_storage->getQuery()->sort('weight')->execute();
-    if (count($link_ids) > 0) {
-      /** @var \Drupal\oe_corporate_blocks\Entity\FooterLinkInterface $link_entity */
-      foreach ($links_storage->loadMultiple($link_ids) as $link_entity) {
-        $cache->addCacheableDependency($link_entity);
-
-        $link = [
-          'href' => $link_entity->getUrl()->toString(),
-          'label' => $link_entity->label(),
-        ];
-
-        if ($link_entity instanceof FooterLinkSocialInterface) {
-          $link['social_network'] = $link_entity->getSocialNetwork();
-        }
-
-        $links[] = $link;
-      }
-    }
-
-    return $links;
   }
 
 }
